requirements.txt
PyPDF2==3.0.1
pytesseract==0.3.10
langchain==0.1.12
langchain-openai==0.0.5
pdf2image==1.17.0
python-dotenv==1.0.1
pdfplumber==0.10.3
nltk==3.8.1
Pillow==10.2.0 